Increase timeout for root device to appear
This change updates the init script to increase the duration for which
the script will wait for the root device to appear. Previously, it would
wait at most 10s, but user reports show that it might take up to 30s for
the root device to actually appear on the UDMP (see #10).