Load native WireGuard kernel module
This change fixes the init script to ignore the WireGuard module shipped
in the Ubiquiti stock firmware. This module is not compatible with
custom kernels. See the following issue:
https://github.com/fabianishere/udm-kernel/issues/13

Currently, the init script will configure modprobe to force-load
Ubiquiti's WireGuard version. This change updates it to ignore that
WireGuard kernel module, in order to use the native module.